Revert "Validate added co-owners i ChangeCoOwnerOfVehicle"
This reverts commit 719277e833dc596179d09c0ce3b319d2142e00f0.
