+++ libs/application/templates/transport-authority/change-co-owner-of-vehicle/src/fields/CoOwner/index.tsx
@@ -17,9 +17,7 @@ import { information } from '../../lib/messages'
 import { OwnerCoOwnersInformation } from '../../shared'
 import { useFormContext } from 'react-hook-form'
 
-export const OwnerCoOwners: FC<React.PropsWithChildren<FieldBaseProps>> = (
-  props,
-) => {
+export const CoOwner: FC<React.PropsWithChildren<FieldBaseProps>> = (props) => {
   const { application, field, setFieldLoadingState, errors } = props
 
   const { formatMessage } = useLocale()
