+++ libs/application/templates/transport-authority/change-co-owner-of-vehicle/src/fields/index.ts
@@ -1,7 +1,6 @@
 export { VehiclesField } from './VehiclesField'
-export { OwnerCoOwners } from './OwnerCoOwners'
-export { OldCoOwnerRepeater } from './OldCoOwnerRepeater'
-export { NewCoOwnerRepeater } from './NewCoOwnerRepeater'
+export { CoOwner } from './CoOwner'
+export { CoOwnerRepeater } from './CoOwnerRepeater'
 export { Review } from './Review'
 export { RejectedConclusion } from './RejectedConclusion'
 export { ValidationErrorMessages } from './ValidationErrorMessages'
