+++ libs/application/templates/transport-authority/change-co-owner-of-vehicle/src/forms/ChangeCoOwnerOfVehicleForm/InformationSection/coOwnerSubSection.ts
@@ -15,14 +15,8 @@ export const coOwnerSubSection = buildSubSection({
       description: information.labels.coOwner.description,
       children: [
         buildCustomField({
-          id: 'ownerCoOwners',
-          component: 'OldCoOwnerRepeater',
-          title: '',
-          doesNotRequireAnswer: true,
-        }),
-        buildCustomField({
           id: 'coOwners',
-          component: 'NewCoOwnerRepeater',
+          component: 'CoOwnerRepeater',
           title: '',
           doesNotRequireAnswer: true,
         }),
