+++ libs/application/templates/transport-authority/change-co-owner-of-vehicle/src/forms/ChangeCoOwnerOfVehicleForm/InformationSection/ownerSubSection.ts
@@ -63,7 +63,7 @@ export const ownerSubSection = buildSubSection({
         buildCustomField({
           id: 'ownerCoOwners',
           title: '',
-          component: 'OwnerCoOwners',
+          component: 'CoOwner',
           doesNotRequireAnswer: true,
         }),
       ],
