+++ libs/application/templates/transport-authority/transfer-of-vehicle-ownership/src/fields/BuyerCoOwnerAndOperatorRepeater/index.tsx
@@ -61,7 +61,9 @@ export const BuyerCoOwnerAndOperatorRepeater: FC<
   }
 
   const getBuyerNationalId = async () => {
-    // Get updated buyer nationalId from answers
+    // Get buyer nationalId from updated answers (cannot use application.answers),
+    // since that is filled out in the same step as this,
+    // but the buyer value is updated in answers right away
     const applicationInfo = await getApplicationInfo({
       variables: {
         input: {
@@ -71,6 +73,7 @@ export const BuyerCoOwnerAndOperatorRepeater: FC<
       },
       fetchPolicy: 'no-cache',
     })
+
     const updatedApplication = applicationInfo?.data?.applicationApplication
 
     const buyerNationalId =
